Fix possible softlock with the rotation block puzzle in the depths; this involved creating a new layer to detect when a block enters a trigger, which will un-solve the room when the player pushes the block all the way to the right of the room, then leaves the room. The room to the left of the puzzle, "End Of Challenge Loop" has a condition combiner which checks if the block was pushed into the invisible trigger on the right side of the puzzle room, and if the player entered the "End of challenge loop" from the left. This un-solving of the room should eliminate the soft lock.
I also changed the puzzle to require the player to grapple to the "final" line the block rests on and rotate it, which prevents the player from just shooting the block across the room. Also to note, the puzzle will not "solve" officially until the player leaves the room. There is one flag which checks if the block is set into the pressure plate, and another flag which completes the room which is only completed after the player leaves the room through the left or bottom.